+++ src/pluginHooks.js
@@ -5,6 +5,8 @@ Handsontable.PluginHooks = {
     afterLoadData: [],
     beforeRender: [],
     afterRender: [],
+    beforeGet: [],
+    beforeSet: [],
     afterGetCellMeta: [],
     afterGetColHeader: [],
     afterGetColWidth: [],
@@ -19,14 +21,10 @@ Handsontable.PluginHooks = {
     this.hooks[hook].unshift(fn);
   },
 
-  run: function (instance, hook, args) {
+  run: function (instance, hook, p1, p2, p3, p4, p5) {
+    //performance considerations - http://jsperf.com/call-vs-apply-for-a-plugin-architecture
     for (var i = 0, ilen = this.hooks[hook].length; i < ilen; i++) {
-      if (args) {
-        this.hooks[hook][i].apply(instance, args);
-      }
-      else {
-        this.hooks[hook][i].call(instance);
-      }
+      this.hooks[hook][i].call(instance, p1, p2, p3, p4, p5);
     }
   }
 };